+++ test/app/components/another_component.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
 class AnotherComponent < ViewComponent::Base
-  def initialize(*); end
 end