+++ test/app/components/editorb_component.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
 class EditorbComponent < ViewComponent::Base
-  def initialize(*); end
 end