+++ test/app/components/inline_component.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 class InlineComponent < ViewComponent::Base
-  def initialize(*); end
-
   def call
     text_field_tag :name
   end