+++ test/app/components/my_component.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
 class MyComponent < ViewComponent::Base
-  def initialize(*); end
 end