+++ test/app/components/path_container_component.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
 class PathContainerComponent < ViewComponent::Base
-  def initialize(*)
-  end
 end