+++ test/app/components/unreferenced_component.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
 class UnreferencedComponent < ViewComponent::Base
-  def initialize(*)
-  end
 end