+++ test/app/components/url_component.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
 class UrlComponent < ViewComponent::Base
-  def initialize(*)
-  end
 end