+++ test/app/components/wrapper_component.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
 
 class WrapperComponent < ViewComponent::Base
-  def initialize(*); end
 end